backend/requirements.txt
flask
flask-cors
python-dotenv
werkzeug
google-generativeai
google-cloud-speech
google-cloud-texttospeech
google-cloud-vision
moviepy
